+++ spec/lib/converter_spec.rb
@@ -6,8 +6,9 @@ module MinitestToRspec
     SPEC_FIXTURES = File.join(__dir__, '..', 'fixtures')
     FIXTURE_DIRS = Dir.glob("#{SPEC_FIXTURES}/*")
 
-    def convert(input)
-      described_class.new.convert(input)
+    def convert(input, options = nil)
+      options ||= { rails_helper: false }
+      described_class.new(options).convert(input)
     end
 
     def input(fixture)
@@ -25,6 +26,12 @@ module MinitestToRspec
           calculated = convert(input(fixture)).strip
           expect(calculated).to eq(expected)
         end
+      end
+
+      it "supports rails_helper option" do
+        expect(
+          convert("require 'test_helper'", rails_helper: true)
+        ).to eq('require("rails_helper")')
       end
     end
   end