+++ spec/lib/processor_spec.rb
@@ -2,7 +2,7 @@ require "spec_helper"
 
 module MinitestToRspec
   RSpec.describe Processor do
-    let(:processor) { described_class.new }
+    let(:processor) { described_class.new(false) }
 
     describe "#process_call" do
       let(:delegate) { Subprocessors::Call }
@@ -11,7 +11,7 @@ module MinitestToRspec
         input = s(:call)
         allow(delegate).to receive(:process).and_call_original
         processor.process(input)
-        expect(delegate).to have_received(:process).with(input)
+        expect(delegate).to have_received(:process).with(input, false)
       end
     end